refactor: create separate packages for `confirm no id` and `abort` screens
Resolves: DCMAW-10691
